add geometry indexed solid color effect/scene

+++ src/Effects/DefaultGeometryShader.h
@@ -14,7 +14,7 @@ public:
     typedef Vertex Output;
 public:
     Triangle<Output> operator()(
-        const Vertex& in0, const Vertex& in1, const Vertex& in2, unsigned int triangle_index) const
+        const Vertex& in0, const Vertex& in1, const Vertex& in2, size_t triangle_index) const
     {
         return { in0, in1, in2 };
     }
